turn table into macro
