+++ test/doggo_test.exs
@@ -5,7 +5,6 @@ defmodule DoggoTest do
   import Doggo.TestHelpers
 
   alias Phoenix.LiveView.JS
-  alias Phoenix.LiveView.LiveStream
 
   describe "alert/1" do
     test "default" do
@@ -4450,153 +4449,6 @@ defmodule DoggoTest do
         """)
 
       assert attribute(html, "button:root", "data-test") == "hello"
-    end
-  end
-
-  describe "table/1" do
-    test "default" do
-      assigns = %{pets: [%{id: 1, name: "George"}]}
-
-      html =
-        parse_heex(~H"""
-        <Doggo.table id="pets" rows={@pets}>
-          <:col :let={p} label="Name"><%= p.name %></:col>
-          <:action :let={p} label="Link">link-to-<%= p.id %></:action>
-        </Doggo.table>
-        """)
-
-      assert attribute(html, "div:root", "class") == "table-container"
-      assert attribute(html, ":root > table", "id") == "pets"
-      assert text(html, "table > thead > tr > th:first-child") == "Name"
-      assert text(html, "table > thead > tr > th:last-child") == "Link"
-      assert attribute(html, "table > tbody", "id") == "pets-tbody"
-      assert text(html, "tbody > tr > td:first-child") == "George"
-      assert text(html, "tbody > tr > td:last-child") == "link-to-1"
-    end
-
-    test "with caption" do
-      assigns = %{pets: [%{id: 1, name: "George"}]}
-
-      html =
-        parse_heex(~H"""
-        <Doggo.table id="pets" rows={@pets} caption="some text">
-          <:col :let={p} label="Name"><%= p.name %></:col>
-        </Doggo.table>
-        """)
-
-      assert text(html, "table > caption") == "some text"
-    end
-
-    test "with col attrs on column" do
-      assigns = %{pets: [%{id: 1, name: "George"}]}
-
-      html =
-        parse_heex(~H"""
-        <Doggo.table id="pets" rows={@pets}>
-          <:col :let={p} label="Name" col_attrs={[style: "width: 20%;"]}>
-            <%= p.name %>
-          </:col>
-          <:action :let={p} label="Link">link-to-<%= p.id %></:action>
-        </Doggo.table>
-        """)
-
-      assert attribute(html, "table > colgroup > col:first-child", "style") ==
-               "width: 20%;"
-    end
-
-    test "with col attrs on action" do
-      assigns = %{pets: [%{id: 1, name: "George"}]}
-
-      html =
-        parse_heex(~H"""
-        <Doggo.table id="pets" rows={@pets}>
-          <:col :let={p} label="Name"><%= p.name %></:col>
-          <:action :let={p} label="Link" col_attrs={[style: "width: 20%;"]}>
-            link-to-<%= p.id %>
-          </:action>
-        </Doggo.table>
-        """)
-
-      assert attribute(html, "table > colgroup > col:last-child", "style") ==
-               "width: 20%;"
-    end
-
-    test "with foot" do
-      assigns = %{pets: [%{id: 1, name: "George"}]}
-
-      html =
-        parse_heex(~H"""
-        <Doggo.table id="pets" rows={@pets}>
-          <:col :let={p} label="Name"><%= p.name %></:col>
-          <:foot>some foot</:foot>
-        </Doggo.table>
-        """)
-
-      assert text(html, "table > tfoot") == "some foot"
-    end
-
-    test "with row id" do
-      assigns = %{pets: [%{id: 1, name: "George"}]}
-
-      html =
-        parse_heex(~H"""
-        <Doggo.table id="pets" rows={@pets} row_id={&"row-#{&1.id}"}>
-          <:col :let={p} label="Name"><%= p.name %></:col>
-        </Doggo.table>
-        """)
-
-      assert attribute(html, "tbody > tr", "id") == "row-1"
-    end
-
-    test "with row click" do
-      assigns = %{pets: [%{id: 1, name: "George"}]}
-
-      html =
-        parse_heex(~H"""
-        <Doggo.table id="pets" rows={@pets} row_click={&"clicked-#{&1.id}"}>
-          <:col :let={p} label="Name"><%= p.name %></:col>
-        </Doggo.table>
-        """)
-
-      assert attribute(html, "tbody td", "phx-click") == "clicked-1"
-    end
-
-    test "with row item" do
-      assigns = %{pets: [%{id: 1, name: "George"}]}
-
-      html =
-        parse_heex(~H"""
-        <Doggo.table id="pets" rows={@pets} row_item={&Map.put(&1, :name, "G")}>
-          <:col :let={p} label="Name"><%= p.name %></:col>
-        </Doggo.table>
-        """)
-
-      assert text(html, "tbody td") == "G"
-    end
-
-    test "with stream" do
-      assigns = %{
-        pets:
-          LiveStream.new(
-            :pets,
-            0,
-            [%{id: 1, name: "George"}, %{id: 2, name: "Mary"}],
-            []
-          )
-      }
-
-      html =
-        parse_heex(~H"""
-        <Doggo.table id="pets" rows={@pets}>
-          <:col :let={{id, p}} label="Name"><%= id %> <%= p.name %></:col>
-        </Doggo.table>
-        """)
-
-      assert attribute(html, "tbody", "phx-update") == "stream"
-      assert attribute(html, "tbody tr:first-child", "id") == "pets-1"
-      assert attribute(html, "tbody tr:last-child", "id") == "pets-2"
-      assert text(html, "tbody tr:first-child td") == "pets-1 George"
-      assert text(html, "tbody tr:last-child td") == "pets-2 Mary"
     end
   end
